MCOL-4802 Removed ByteStream methods for bool and add some logging in…

+++ dbcon/joblist/batchprimitiveprocessor-jl.cpp
@@ -1127,7 +1127,7 @@ void BatchPrimitiveProcessorJL::createBPP(ByteStream& bs) const
                     // MCOL-4173 Notify PP if smallSide and largeSide have different column widths
                     // and send smallSide RG to PP.
                     bool joinHasSkewedKeyColumn = tJoiners[i]->joinHasSkewedKeyColumn();
-                    bs << joinHasSkewedKeyColumn;
+                    bs << (uint8_t) joinHasSkewedKeyColumn;
                     if (!smallSideRGSent && joinHasSkewedKeyColumn)
                     {
                         idbassert(!smallSideRGs.empty());
